// app/frontend/src/shared/schemas/matchesSchemas.ts
export type TournamentData = {
	pairs: { player1: string, player2: string }[];
	results: (number | null)[];
	round: number;
};